+++ tests/test_checker/test_noqa.py
@@ -150,6 +150,7 @@ def test_noqa_fixture_disabled(absolute_path, all_violations):
         'Z461': 1,
         'Z462': 1,
         'Z463': 1,
+        'Z464': 1,
     }
 
     process = subprocess.Popen(