+++ tests/test_visitors/test_ast/test_exceptions/test_try_finally.py
@@ -3,63 +3,77 @@
 import pytest
 
 from wemake_python_styleguide.violations.best_practices import (
-    BaseExceptionViolation,
+    RedundantFinallyViolation,
 )
-from wemake_python_styleguide.visitors.ast.keywords import (
+from wemake_python_styleguide.visitors.ast.exceptions import (
     WrongTryExceptVisitor,
 )
 
-use_base_exception = """
-try:
-    ...
-except BaseException:
-    ...
-"""
-
-use_except_exception = """
-try:
-    ...
-except Exception:
-    ...
-"""
-
-use_bare_except = """
+right_try_example = """
 try:
     ...
 except:
     ...
 """
 
+wrong_try_example = """
+try:
+    ...
+finally:
+    ...
+"""
+
+check_finally_with_except = """
+try:
+    ...
+except:
+    ...
+finally:
+    ...
+"""
+
+check_finally_with_except_else = """
+try:
+    ...
+except Exception:
+    ...
+else:
+    ...
+finally:
+    ...
+"""
+
 
 @pytest.mark.parametrize('code', [
-    use_base_exception,
+    wrong_try_example,
 ])
-def test_use_base_exception(
+def test_wrong_finally(
     assert_errors,
     parse_ast_tree,
     code,
     default_options,
 ):
-    """Testing that `except BaseException:` is restricted."""
+    """Violations are raised when finally without except in try block."""
     tree = parse_ast_tree(code)
 
     visitor = WrongTryExceptVisitor(default_options, tree=tree)
     visitor.run()
 
-    assert_errors(visitor, [BaseExceptionViolation])
+    assert_errors(visitor, [RedundantFinallyViolation])
 
 
 @pytest.mark.parametrize('code', [
-    use_except_exception,
-    use_bare_except,
+    right_try_example,
+    check_finally_with_except,
+    check_finally_with_except_else,
 ])
-def test_use_exception(
+def test_correct_finally(
     assert_errors,
     parse_ast_tree,
     code,
     default_options,
 ):
-    """Testing that `except Exception:` and `except:` are allowed."""
+    """Violations are not raised when finally with except in try block."""
     tree = parse_ast_tree(code)
 
     visitor = WrongTryExceptVisitor(default_options, tree=tree)