+++ tests/test_visitors/test_ast/test_exceptions/test_useless_except_case.py
@@ -5,7 +5,7 @@ import pytest
 from wemake_python_styleguide.violations.consistency import (
     UselessExceptCaseViolation,
 )
-from wemake_python_styleguide.visitors.ast.keywords import (
+from wemake_python_styleguide.visitors.ast.exceptions import (
     WrongTryExceptVisitor,
 )