+++ CHANGELOG.md
@@ -20,11 +20,12 @@ We used to have incremental versioning before `0.1.0`.
 - Forbids to use direct magic attributes access
 - Forbids to use negated conditions
 - Forbids to use too many `# pragma: no cover` comments
+- Forbids to use nested `try` blocks
 
 ### Bugfixes
 
 - Fixes problems with empty lines after magic comments, see [#492](https://github.com/wemake-services/wemake-python-styleguide/issues/492)
-- Fixes error message for `del` keyword
+- Fixes error message for `del` keyword: it is now just `'del'` not `'delete'`
 
 ### Misc